+++ codegen/src/module.rs
@@ -1,12 +1,8 @@
 use quote::{quote, ToTokens};
 use syn::{parse::Parse, parse::ParseStream};
 
-#[cfg(no_std)]
-use core::mem;
-#[cfg(not(no_std))]
-use std::mem;
-
 use std::borrow::Cow;
+use std::mem;
 
 use crate::attrs::{AttrItem, ExportInfo, ExportScope, ExportedParams};
 use crate::function::ExportedFn;