+++ tests/debugging.rs
@@ -58,7 +58,7 @@ fn test_debugger_state() {
             println!("Current state = {}", context.global_runtime_state().debugger().state());
 
             // Modify state
-            let mut state = context.global_runtime_state_mut().debugger_mut().state_mut().write_lock::<Map>().unwrap();
+            let mut state = context.global_runtime_state_mut().debugger_mut().state_mut().as_map_ref_mut().unwrap();
             let hello = state.get("hello").unwrap().as_int().unwrap();
             state.insert("hello".into(), (hello + 1).into());
             state.insert("foo".into(), true.into());